+++ NEWS.rst
@@ -4,9 +4,10 @@ News for sak
 This file lists the major changes between versions. For a more detailed list
 of every change, see the Git log.
 
-Latest
-------
-* tbd
+Latest (feature release)
+------------------------
+* Added the object registry, which allows management and access to
+  object factories.
 
 4.2.2
 -----